User locked down

+++ src/resolvers/Mutation.js
@@ -1,6 +1,6 @@
 import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
-
+import { getUserId } from '../util';
 export default {
   async createUser(parent, args, { prisma }, info) {
     const passwordLen = args.data.password && args.data.password.length > 7;
@@ -31,33 +31,36 @@ export default {
       throw new Error('Unable to login');
     }
     return {
-      token: jwt.sign({ id: user.id }, 'thisismysecret'),
+      token: jwt.sign({ userId: user.id }, 'thisismysecret'),
       user,
     };
   },
-  async deleteUser(parent, args, { prisma }, info) {
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.deleteUser(
       {
         where: {
-          id: args.id,
+          id: userId,
         },
       },
       info
     );
   },
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id,
+          id: userId,
         },
         data: args.data,
       },
       info
     );
   },
-  async createPost(parent, { data }, { prisma }, info) {
+  async createPost(parent, { data }, { prisma, request }, info) {
     const { title, body, published } = data;
+    const userId = getUserId(request);
     return prisma.mutation.createPost(
       {
         data: {
@@ -66,7 +69,7 @@ export default {
           published,
           author: {
             connect: {
-              id: data.author,
+              id: userId,
             },
           },
         },
@@ -85,7 +88,17 @@ export default {
       info
     );
   },
-  async deletePost(parent, { id }, { prisma }, info) {
+  async deletePost(parent, { id }, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!postExists) {
+      throw new Error('Unable to delete post');
+    }
     return prisma.mutation.deletePost(
       {
         where: {
